+++ app/crates/kartoffels-world/src/bots/systems/kill.rs
@@ -36,7 +36,7 @@ pub fn kill(
 
         let decision = if !killed.oneshot
             && policy.auto_respawn
-            && bots.queued.len() < policy.max_queued_bots
+            && bots.queued.len() < policy.max_queued_bots as usize
         {
             Decision::Requeue
         } else {
